backend/requirements.txt
appdirs==1.4.3
attrs==19.1.0
cached-property==1.5.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.13
django-cors-headers==2.5.2
django-filter==2.1.0
django-guardian==1.4.1
django-sendfile==0.3.11
django-stubs==1.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
html2text==2014.12.29
idna==2.8
isodate==0.6.0
lxml==4.4.1
Markdown==3.1
mypy==0.720
mypy-extensions==0.4.1
mysqlclient==1.4.2.post1
numpy==1.17.2
pandas==0.25.1
Pillow==6.2.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.6
xlrd==1.2.0
zeep==3.4.0
